ListView commit
I've made it to where the restaurants are now displayed in a listView
item. The Italian food button temporarily redirects to an example of
this. I have also made an adapter to handle a specific method of input
So rather than having to manually punch in each restaurant name and
address we can simply add them to an arraylist and it will automatically
populate the listview with that data. The button made by the listview
should also correctly redirect the google maps to the address of the
restaurant. I have also added a drop down menu that should pass the city
through the intent. I figure we can create separate ArrayLists for each
city for now and the pages will handle each one.